requirements.txt
asyncio==3.4.3
asyncio-nats-client==0.10.0
asyncqt==0.7.0
helpdev==0.6.10
importlib-metadata==1.2.0
more-itertools==8.0.2
numpy==1.17.4
opencv-contrib-python==4.1.2.30
psutil==5.6.7
PyQt5==5.13.2
PyQt5-sip==12.7.0
PyQt5-stubs==5.13.1.3
zipp==0.6.0
